+++ R/reexport.R
@@ -1,0 +1,75 @@
+#' @aliases reexports
+
+# magrittr ---------------------------------------------------------------------
+#' @export
+#' @importFrom magrittr %>%
+magrittr::`%>%`
+
+# tibble -----------------------------------------------------------------------
+#' @export
+#' @importFrom tibble as_tibble
+tibble::as_tibble
+
+# knitr ------------------------------------------------------------------------
+#' @export
+#' @importFrom knitr knit_print
+knitr::knit_print
+
+# dplyr ------------------------------------------------------------------------
+#' @importFrom dplyr vars
+#' @export
+dplyr::vars
+
+# tidyselect -------------------------------------------------------------------
+#' @importFrom tidyselect starts_with
+#' @export
+tidyselect::starts_with
+
+#' @importFrom tidyselect ends_with
+#' @export
+tidyselect::ends_with
+
+#' @importFrom tidyselect contains
+#' @export
+tidyselect::contains
+
+#' @importFrom tidyselect matches
+#' @export
+tidyselect::matches
+
+#' @importFrom tidyselect num_range
+#' @export
+tidyselect::num_range
+
+#' @importFrom tidyselect all_of
+#' @export
+tidyselect::all_of
+
+#' @importFrom tidyselect any_of
+#' @export
+tidyselect::any_of
+
+#' @importFrom tidyselect everything
+#' @export
+tidyselect::everything
+
+#' @importFrom tidyselect last_col
+#' @export
+tidyselect::last_col
+
+#' @importFrom tidyselect one_of
+#' @export
+tidyselect::one_of
+
+# bstfun -----------------------------------------------------------------------
+#' @importFrom bstfun tbl_ancova
+#' @export
+bstfun::tbl_ancova
+
+#' @importFrom bstfun tbl_cross
+#' @export
+bstfun::tbl_cross
+
+#' @importFrom bstfun fix_database_error
+#' @export
+bstfun::fix_database_error