+++ tests/testthat/test-auc_density.R
@@ -1,0 +1,11 @@
+context("auc_density")
+
+test_that("No errors/warnings with standard use", {
+
+  expect_error(
+    auc_density(density = dbeta, shape1 = 1, shape2 = 1),
+    NA
+  )
+
+})
+