+++ tests/testthat/test-auc_histogram.R
@@ -1,0 +1,10 @@
+context("auc_histogram")
+
+test_that("No errors/warnings with standard use", {
+
+  expect_error(
+    runif(10000) %>% hist(breaks = 250) %>% auc_histogram(),
+    NA
+  )
+
+})