+++ tests/testthat/test-count_map.R
@@ -10,3 +10,25 @@ test_that("No Errors/Warning with standard use", {
     NA
   )
 })
+
+test_that("Works for a single vector", {
+
+  expect_error(
+  trial %>% count_map(c("stage", "grade")),
+  NA
+  )
+})
+
+test_that("Gives error for incorrect variable names", {
+
+  expect_error(
+    trial %>% count_map(list(c("tstage", "grade"))),
+    "*"
+  )
+
+  expect_error(
+    trial %>% dplyr::mutate(..n.. = 1) %>% count_map(list(c("..n..", "grade"))),
+    "*"
+  )
+
+})